server/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
catboost>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
kaleido>=0.2.1
scipy>=1.9.0
statsmodels>=0.13.0
shap>=0.41.0
lime>=0.2.0
tqdm>=4.64.0
debugpy>=1.6.0
dill>=0.3.6
requests>=2.28.0
umap-learn>=0.5.3